requirements.txt
Django==1.8.3
django-easymoney==0.7.2
django-filter==1.0.2
